feat(ihev2): nots 2
Refs: #1667
